+++ src/app/showcase-espacemembre/index.ts
@@ -1,0 +1,3 @@
+export * from './showcase-espacemembre.component';
+export * from './students-home-page';
+
