+++ src/app/showcase-espacemembre/showcase-espacemembre.component.html
@@ -1,0 +1,6 @@
+<span id="top"></span>
+<div id="showcase-espacemembre">
+  <main class="content">
+    <router-outlet (deactivate)="onDeactivate()"></router-outlet>
+  </main>
+</div>
