+++ src/app/showcase-espacemembre/students-home-page/index.ts
@@ -1,0 +1,1 @@
+export * from './students-home-page.component';
