+++ src/app/showcase-espacemembre/students-home-page/students-home-page.component.html
@@ -1,0 +1,2 @@
+<div>Hello World</div>
+<router-outlet></router-outlet>
