+++ src/app/showcase-espacemembre/students-home-page/students-home-page.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute,Router } from '@angular/router';
-import { Link } from 'app/shared';
+
 @Component({
   selector: 'students-home-page',
   templateUrl: './students-home-page.component.html',
@@ -8,9 +8,6 @@ import { Link } from 'app/shared';
 })
 export class StudentsHomePageComponent implements OnInit {
 
-  navLinks: Link[] = [
-    {href: '/sorties', text: 'Sorties'}, 
-  ];
   constructor(private router: Router) { }
 
   ngOnInit() {
