--- src/app/students-home-page/students-home-page.component.html
@@ -1,2 +1,0 @@
-<app-navbar [links]="navLinks"></app-navbar>
-<router-outlet></router-outlet>
