--- src/app/students-home-page/students-home-page.module.ts
@@ -1,31 +1,0 @@
-import { NgModule } from '@angular/core';
-
-// Modules
-import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
-import { RouterModule } from '@angular/router';
-import { AgmCoreModule } from '@agm/core';
-import { MomentModule } from 'ngx-moment';
-import { CoreModule } from 'app/core';
-import { SharedModule } from 'app/shared';
-import { StudentsHomePageRoutingModule } from './students-home-page-routing.module';
-
-// Components
-import { StudentsHomePageComponent } from './students-home-page.component';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    MomentModule,
-    RouterModule,
-    AgmCoreModule,
-    CoreModule,
-    SharedModule,
-    StudentsHomePageRoutingModule
-  ],
-  declarations: [
-    StudentsHomePageComponent,
-  ],
-})
-export class StudentsHomePageModule { }
