+++ src/index.html
@@ -1,6 +1,8 @@
+
 <!doctype html>
 <html>
 <head>
+<!-- Page qui sert à afficher la roue tournante "loading"-->
   <meta charset="utf-8">
   <title>OSER</title>
   <base href="/">
